refactor(dashboard): deduplicate store normalizers into store-normalizers.ts
Move 4 shared normalizer functions to store-normalizers.ts SSOT:
- normalizeAttentionItem (room-truth-store + mission-store → store-normalizers)
- normalizeRecommendedAction (room-truth-store + mission-store → store-normalizers)
- normalizeExecutionSummary (room-truth-store → import from store-normalizers)
- normalizeExecutionQueueItem (room-truth-store → import from store-normalizers)

normalizeServerStatus kept local in room-truth-store (different signature from
store-normalizers version — simple field-pick vs spread-and-overlay).

Net: -77 lines (116 removed, 39 added).
